+++ service-1/application/models.py
@@ -1,4 +1,4 @@
-from app import db
+from application import db
 
 
 class character(db.Model):
