+++ service-1/create.py
@@ -1,5 +1,5 @@
-from app import db
-from models import character
+from application import db
+from application.models import character
 
-
+# Creating my database table to store my data
 db.create_all()
