--- service-1/forms.py
@@ -1,2 +1,0 @@
-from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField
