+++ service-3/application/__init__.py
@@ -1,0 +1,4 @@
+from application import routes
+from flask import Flask, request
+import requests
+from application import routes
